requirements.txt
blinker==1.9
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
Werkzeug==3.1.3
zipp==3.21.0
requests==2.32.3
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
json5==0.10.0
psycopg2-binary==2.9.10
pytest==8.4.1
pytest-docker==3.2.3
docker==7.1.0
pyyaml==6.0.2
